Take care of stuck jobs
Previously it was possible for an Awk job to keep lingering on, for
example if it was stuck in an infinite loop. This commit takes care of
the issue by making sure there is only ever one job running at a time.
If a new job needs to be started, the old one is terminated.